test(#zimic): split fetch client body tests into dedicated files
