+++ packages/zimic/src/fetch/__tests__/FetchClient.bodies.json.test.ts
@@ -1,0 +1,324 @@
+import { describe, expect, expectTypeOf, it } from 'vitest';
+
+import { HttpSchema, StrictHeaders } from '@/http';
+import { joinURL } from '@/utils/urls';
+import { usingHttpInterceptor } from '@tests/utils/interceptors';
+
+import createFetch from '../factory';
+import { FetchRequest, FetchResponse } from '../types/requests';
+
+describe('FetchClient (node) > Bodies > JSON', () => {
+  const baseURL = 'http://localhost:3000';
+
+  interface User {
+    name: string;
+  }
+
+  const users: User[] = [{ name: 'User 1' }, { name: 'User 2' }];
+
+  it('should support requests and responses with a JSON body', async () => {
+    type Schema = HttpSchema<{
+      '/users': {
+        POST: {
+          request: {
+            headers: { 'content-type': 'application/json' };
+            body: User;
+          };
+          response: {
+            201: { body: User };
+          };
+        };
+      };
+    }>;
+
+    await usingHttpInterceptor<Schema>({ type: 'local', baseURL }, async (interceptor) => {
+      await interceptor
+        .post('/users')
+        .with({ body: users[0] })
+        .respond({
+          status: 201,
+          body: users[0],
+        })
+        .times(1);
+
+      const fetch = createFetch<Schema>({ baseURL });
+
+      const response = await fetch('/users', {
+        method: 'POST',
+        headers: { 'content-type': 'application/json' },
+        body: JSON.stringify(users[0]),
+      });
+
+      expect(response.status).toBe(201);
+      expect(await response.json()).toEqual(users[0]);
+
+      expect(response).toBeInstanceOf(Response);
+      expectTypeOf(response satisfies Response).toEqualTypeOf<
+        FetchResponse<'/users', 'POST', Schema['/users']['POST']>
+      >();
+
+      expect(response.url).toBe(joinURL(baseURL, '/users'));
+
+      expect(response.headers).toBeInstanceOf(Headers);
+      expectTypeOf(response.headers).toEqualTypeOf<StrictHeaders<never>>();
+
+      expectTypeOf(response.json).toEqualTypeOf<() => Promise<User>>();
+      expectTypeOf(response.text).toEqualTypeOf<() => Promise<string>>();
+      // eslint-disable-next-line @typescript-eslint/unbound-method
+      expectTypeOf(response.arrayBuffer).toEqualTypeOf<() => Promise<ArrayBuffer>>();
+      expectTypeOf(response.formData).toEqualTypeOf<() => Promise<FormData>>();
+      expectTypeOf(response.clone).toEqualTypeOf<() => typeof response>();
+      expectTypeOf(response.error).toEqualTypeOf<null>();
+
+      expect(response.request).toBeInstanceOf(Request);
+      expectTypeOf(response.request satisfies Request).toEqualTypeOf<
+        FetchRequest<'/users', 'POST', Schema['/users']['POST']>
+      >();
+
+      expect(response.request.url).toBe(joinURL(baseURL, '/users'));
+
+      expect(response.request.path).toBe('/users');
+      expectTypeOf(response.request.path).toEqualTypeOf<'/users'>();
+
+      expect(response.request.method).toBe('POST');
+      expectTypeOf(response.request.method).toEqualTypeOf<'POST'>();
+
+      expect(response.request.headers).toBeInstanceOf(Headers);
+      expectTypeOf(response.request.headers).toEqualTypeOf<StrictHeaders<{ 'content-type': 'application/json' }>>();
+
+      expectTypeOf(response.request.json).toEqualTypeOf<() => Promise<User>>();
+      expect(await response.request.json()).toEqual(users[0]);
+      expectTypeOf(response.request.text).toEqualTypeOf<() => Promise<string>>();
+      expectTypeOf(response.request.formData).toEqualTypeOf<() => Promise<FormData>>();
+      expectTypeOf(response.request.clone).toEqualTypeOf<() => typeof response.request>();
+    });
+  });
+
+  it('should support requests and responses with a number as a JSON body', async () => {
+    type Schema = HttpSchema<{
+      '/users': {
+        POST: {
+          request: {
+            headers: { 'content-type': 'application/json' };
+            body: number;
+          };
+          response: {
+            201: { body: number };
+          };
+        };
+      };
+    }>;
+
+    await usingHttpInterceptor<Schema>({ type: 'local', baseURL }, async (interceptor) => {
+      await interceptor
+        .post('/users')
+        .with({ body: 1 })
+        .respond({
+          status: 201,
+          body: 2,
+        })
+        .times(1);
+
+      const fetch = createFetch<Schema>({ baseURL });
+
+      const response = await fetch('/users', {
+        method: 'POST',
+        headers: { 'content-type': 'application/json' },
+        body: JSON.stringify(1),
+      });
+
+      expect(response.status).toBe(201);
+      expect(await response.json()).toEqual(2);
+
+      expect(response).toBeInstanceOf(Response);
+      expectTypeOf(response satisfies Response).toEqualTypeOf<
+        FetchResponse<'/users', 'POST', Schema['/users']['POST']>
+      >();
+
+      expect(response.url).toBe(joinURL(baseURL, '/users'));
+
+      expect(response.headers).toBeInstanceOf(Headers);
+      expectTypeOf(response.headers).toEqualTypeOf<StrictHeaders<never>>();
+
+      expectTypeOf(response.json).toEqualTypeOf<() => Promise<number>>();
+      expectTypeOf(response.text).toEqualTypeOf<() => Promise<string>>();
+      // eslint-disable-next-line @typescript-eslint/unbound-method
+      expectTypeOf(response.arrayBuffer).toEqualTypeOf<() => Promise<ArrayBuffer>>();
+      expectTypeOf(response.formData).toEqualTypeOf<() => Promise<FormData>>();
+      expectTypeOf(response.clone).toEqualTypeOf<() => typeof response>();
+      expectTypeOf(response.error).toEqualTypeOf<null>();
+
+      expect(response.request).toBeInstanceOf(Request);
+      expectTypeOf(response.request satisfies Request).toEqualTypeOf<
+        FetchRequest<'/users', 'POST', Schema['/users']['POST']>
+      >();
+
+      expect(response.request.url).toBe(joinURL(baseURL, '/users'));
+
+      expect(response.request.path).toBe('/users');
+      expectTypeOf(response.request.path).toEqualTypeOf<'/users'>();
+
+      expect(response.request.method).toBe('POST');
+      expectTypeOf(response.request.method).toEqualTypeOf<'POST'>();
+
+      expect(response.request.headers).toBeInstanceOf(Headers);
+      expectTypeOf(response.request.headers).toEqualTypeOf<StrictHeaders<{ 'content-type': 'application/json' }>>();
+
+      expectTypeOf(response.request.json).toEqualTypeOf<() => Promise<number>>();
+      expect(await response.request.json()).toEqual(1);
+      expectTypeOf(response.request.text).toEqualTypeOf<() => Promise<string>>();
+      expectTypeOf(response.request.formData).toEqualTypeOf<() => Promise<FormData>>();
+      expectTypeOf(response.request.clone).toEqualTypeOf<() => typeof response.request>();
+    });
+  });
+
+  it('should support requests and responses with a boolean as a JSON body', async () => {
+    type Schema = HttpSchema<{
+      '/users': {
+        POST: {
+          request: {
+            headers: { 'content-type': 'application/json' };
+            body: boolean;
+          };
+          response: {
+            201: { body: boolean };
+          };
+        };
+      };
+    }>;
+
+    await usingHttpInterceptor<Schema>({ type: 'local', baseURL }, async (interceptor) => {
+      await interceptor
+        .post('/users')
+        .with({ body: false })
+        .respond({
+          status: 201,
+          body: true,
+        })
+        .times(1);
+
+      const fetch = createFetch<Schema>({ baseURL });
+
+      const response = await fetch('/users', {
+        method: 'POST',
+        headers: { 'content-type': 'application/json' },
+        body: JSON.stringify(false),
+      });
+
+      expect(response.status).toBe(201);
+      expect(await response.json()).toEqual(true);
+
+      expect(response).toBeInstanceOf(Response);
+      expectTypeOf(response satisfies Response).toEqualTypeOf<
+        FetchResponse<'/users', 'POST', Schema['/users']['POST']>
+      >();
+
+      expect(response.url).toBe(joinURL(baseURL, '/users'));
+
+      expect(response.headers).toBeInstanceOf(Headers);
+      expectTypeOf(response.headers).toEqualTypeOf<StrictHeaders<never>>();
+
+      expectTypeOf(response.json).toEqualTypeOf<() => Promise<boolean>>();
+      expectTypeOf(response.text).toEqualTypeOf<() => Promise<string>>();
+      // eslint-disable-next-line @typescript-eslint/unbound-method
+      expectTypeOf(response.arrayBuffer).toEqualTypeOf<() => Promise<ArrayBuffer>>();
+      expectTypeOf(response.formData).toEqualTypeOf<() => Promise<FormData>>();
+      expectTypeOf(response.clone).toEqualTypeOf<() => typeof response>();
+      expectTypeOf(response.error).toEqualTypeOf<null>();
+
+      expect(response.request).toBeInstanceOf(Request);
+      expectTypeOf(response.request satisfies Request).toEqualTypeOf<
+        FetchRequest<'/users', 'POST', Schema['/users']['POST']>
+      >();
+
+      expect(response.request.url).toBe(joinURL(baseURL, '/users'));
+
+      expect(response.request.path).toBe('/users');
+      expectTypeOf(response.request.path).toEqualTypeOf<'/users'>();
+
+      expect(response.request.method).toBe('POST');
+      expectTypeOf(response.request.method).toEqualTypeOf<'POST'>();
+
+      expect(response.request.headers).toBeInstanceOf(Headers);
+      expectTypeOf(response.request.headers).toEqualTypeOf<StrictHeaders<{ 'content-type': 'application/json' }>>();
+
+      expectTypeOf(response.request.json).toEqualTypeOf<() => Promise<boolean>>();
+      expect(await response.request.json()).toEqual(false);
+      expectTypeOf(response.request.text).toEqualTypeOf<() => Promise<string>>();
+      expectTypeOf(response.request.formData).toEqualTypeOf<() => Promise<FormData>>();
+      expectTypeOf(response.request.clone).toEqualTypeOf<() => typeof response.request>();
+    });
+  });
+
+  it('should consider request with empty JSON bodies as null', async () => {
+    type Schema = HttpSchema<{
+      '/users': {
+        POST: {
+          request: {};
+          response: { 201: {} };
+        };
+      };
+    }>;
+
+    await usingHttpInterceptor<Schema>({ type: 'local', baseURL }, async (interceptor) => {
+      await interceptor
+        .post('/users')
+        .with({ body: null })
+        .respond({
+          status: 201,
+          body: null,
+        })
+        .times(1);
+
+      const fetch = createFetch<Schema>({ baseURL });
+
+      const response = await fetch('/users', {
+        method: 'POST',
+        body: null,
+      });
+
+      expect(response.status).toBe(201);
+      expect(await response.json()).toEqual(null);
+
+      expect(response).toBeInstanceOf(Response);
+      expectTypeOf(response satisfies Response).toEqualTypeOf<
+        FetchResponse<'/users', 'POST', Schema['/users']['POST']>
+      >();
+
+      expect(response.url).toBe(joinURL(baseURL, '/users'));
+
+      expect(response.headers).toBeInstanceOf(Headers);
+      expectTypeOf(response.headers).toEqualTypeOf<StrictHeaders<never>>();
+
+      expectTypeOf(response.json).toEqualTypeOf<() => Promise<null>>();
+      expectTypeOf(response.text).toEqualTypeOf<() => Promise<string>>();
+      // eslint-disable-next-line @typescript-eslint/unbound-method
+      expectTypeOf(response.arrayBuffer).toEqualTypeOf<() => Promise<ArrayBuffer>>();
+      expectTypeOf(response.formData).toEqualTypeOf<() => Promise<FormData>>();
+      expectTypeOf(response.clone).toEqualTypeOf<() => typeof response>();
+      expectTypeOf(response.error).toEqualTypeOf<null>();
+
+      expect(response.request).toBeInstanceOf(Request);
+      expectTypeOf(response.request satisfies Request).toEqualTypeOf<
+        FetchRequest<'/users', 'POST', Schema['/users']['POST']>
+      >();
+
+      expect(response.request.url).toBe(joinURL(baseURL, '/users'));
+
+      expect(response.request.path).toBe('/users');
+      expectTypeOf(response.request.path).toEqualTypeOf<'/users'>();
+
+      expect(response.request.method).toBe('POST');
+      expectTypeOf(response.request.method).toEqualTypeOf<'POST'>();
+
+      expect(response.request.headers).toBeInstanceOf(Headers);
+      expectTypeOf(response.request.headers).toEqualTypeOf<StrictHeaders<never>>();
+
+      expectTypeOf(response.request.json).toEqualTypeOf<() => Promise<null>>();
+      expect(await response.request.text()).toEqual('');
+      expectTypeOf(response.request.text).toEqualTypeOf<() => Promise<string>>();
+      expectTypeOf(response.request.formData).toEqualTypeOf<() => Promise<FormData>>();
+      expectTypeOf(response.request.clone).toEqualTypeOf<() => typeof response.request>();
+    });
+  });
+});
